+++ glean/src/platform/qt/index.ts
@@ -4,14 +4,14 @@
 
 // Qt does not have its implementations yet, we use pieces of
 // the `TestPlatform` so that the sample will still work.
-import uploader from "./uploader.js";
 import type Platform from "../index.js";
-import TestPlatform from "../test/index.js";
+import uploader from "./uploader.js";
 import info from "./platform_info.js";
+import Storage from "./storage.js";
 
 const QtPlatform: Platform = {
-  Storage: TestPlatform.Storage,
-  uploader: uploader,
+  Storage,
+  uploader,
   info,
   name: "Qt"
 };